Update frequency_anchors.R
incorporated the comments Grace left on the last commit. I made "starting_frequency" to be the default but i can change that if we think its better to leave it for the user to choose without a default or if we would like the other option to be the default.